requirements.txt
PyMuPDF>=1.23.0
Pillow>=10.0.0
mobi>=0.3.3

